Updated transaction-failed-modal to new Modal version and renamed the files
